+++ example/android/app/src/main/kotlin/com/dhiabechattaoui/flutter_secure_storage_plus_example/MainActivity.kt
@@ -1,5 +1,5 @@
 package com.dhiabechattaoui.flutter_secure_storage_plus_example
 
-import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.android.FlutterFragmentActivity
 
-class MainActivity : FlutterActivity()
+class MainActivity : FlutterFragmentActivity()